Migrate c++ notes to wiki, fix cast
A longer essay than is probably appropriate to be putting in the source
was needed to be started for another file of pending C++ verification
code.  Since it is auxiliary checking anyway, and writing long essays
in source code is suboptimal, the existing comments from another file
were taken to start the wiki page.

This also reverts a case of not using the "casts for the masses" cast
due to an issue with building non-C constructs with the casts (e.g.
C++11 classes with various types of overloading).  The issue with
the cast has since been resolved, so it's okay to use.