Move tracing and path evaluation into files
The %c-do.c file is one of the biggest offenders of "long file", so
this takes out two categories of routines from it.

One is to move the low-level tracing routines into %d-trace.c.  These
may wind up being supplanted entirely by the debugger, though some
kind of low-level tracing will probably remain in the system.

The other is to separate out path processing functions into %c-path.c.

While doing the change, this also gets rid of the numbered string
function usage for RS_TRACE.  The numbered string IDs might have a
use in theory...for customization or internationalization.  But in
practice--especially for the tracing functions--it's just a holdup to
knowing what you are reading, and some format strings had indeed
gotten out of sync so the parameters were wrong.  If it's low-level
enough to be using raw C strings and not formatting blocks in Rebol,
it's probably low level enough where hardcoding the string is okay.