%d-legacy.c, %c-signal.c, %c-eval.c
Pursuant to the goal of making %c-do.c clearer and more streamlined,
this migrates code out of it and reorders it to get the evaluator
loop itself isolated into its own file as %c-eval.c.  %c-do.c is
thus the entry points to the convenience wrapper routines.

Functions that are only in the debug build to support the legacy mode
are moved into %d-legacy.c.  While previously they had been hidden in
the file and made static so as not to be seen, it's better for them
to be visible to the compiler for linking but not in the way of the
mainline code.

There was only one routine fitting to go in %c-signal.c at the moment,
but there may be more, and it keeps it out of %c-do.c

The debug prep and per-step checks were placed before the evaluator
code itself, but since they are broken out completely they can be
even more separated by having them at the end of the evaluator file.