Step toward clarifying Path Dispatch
This cleans up the historically confusing path dispatch code some,
while shedding some light on the stranger hooks in it.

Variables in the PVS and in function parameters are made const correct,
to more clearly see what is changing.  Names are also made more obvious,
most notably that `pvs->path` was *not* a PATH!, but rather an item
in the traversal of path elements.  So it is now `pvs->item`, which
hopefully helps understanding a little.

Removing several PE_XXX values related to errors simplifies the path
dispatch hook by allowing those hooks to call fail() directly, yet
letting them re-use common error logic by making errors that are
parameterized by the PVS.

The results of investigating why path dispatch isn't simpler are also
added as comments in the code.  More waves of improvements are needed,
but this hopefully gets things started.