Field for relative binding in ANY-SERIES!
This commit does an exploratory move of a copy of the binding target
information from words into ANY-SERIES!, so that an ANY-ARRAY! can
carry along the specific context to which relative words in its
deep copy should be bound.

In doing so, it rearranges the fields in Any_Series's payload.  This
caused problems in the past because some routines were assuming that
Any_Series shared a layout with objects (now Any_Context).  In the
intervening time it appears the assumption is not causing a problem
within the known tests.

So even though this commit does not contain usages of the field, it
puts it in place to get some field testing to make sure that its
existence does not cause problems.  Also, now that the RELATIVE bit
is known to apply not just to words, it is simply VALUE_FLAG_RELATIVE.