Fold Do_Commands() state passing into normal DO
Although it is fully possible to invoke a "COMMAND!" as an ordinary
function call when DOing a block, there was a special kind of DO for
commands known as Do_Commands().  The documentation suggested it would
be faster than ordinary evaluation.

(It would not be faster by much, if it was at all.  So it didn't really
do that.  What it did do was duplicate code, introduce new weird rules
in evaluation, and more or less rule out good error reporting or
possible debugger integration with the main evaluator.)

It had one feature that DO did not have: the ability to thread through
a void*'s worth of state from the caller of Do_Commands() to the
commands themselves.  There will need to be better ways of doing that,
but for now this simplifies the feature to use the normal evaluator to
retain service for the two calls to RL_Do_Commands() in R3-View.