Get rid of RL_Callback (Note: not FFI CALLBACK!)
The R3-Alpha extension mechanism had a callback mechanism that is not
being used in R3-View.  It may well not have been in working order.
What it had been gradually turned into was just a call into the DO
evaluator for RXIARG arrays.

This removes it in order to have less total code to maintain for the
moment.  What the code is doing was fairly easy to recreate (and has
nearly identical code in Redo_Func_Throws).