Make_Typeset => Update_Typeset_Bits
Make_Typeset actually would only write the bits inside of the value,
and did not set the header.  This does a renaming to clarify that
point and solidifies the contract by making the callers set the
typeset header before calling so it can assert it's a typeset.

Commented and general tidying, including removal of an unused function
called Find_Typeset.