Simplify WORD_FLAG_BOUND + VALUE_FLAG_RELATIVE
The three binding states for a word are unbound, bound specifically,
and bound relatively.  Previously the relative flag and specific flag
were set separately so they could be tested for individually, then
to know if something was bound both flags were tested.

This simplifies matters in light of the new value-global flag for
VALUE_FLAG_RELATIVE, that applies also to ANY-ARRAY!.  Now relative
values also have a generic bound bit set.  This means only one bit
needs to be checked for whether a word is bound or not, while a test
for a specific binding takes two tests (bound and not relative).

In the micro-optimizing future, it may be that there is a better way
to do this test, but this simplifies the matters for readability at
this time.