Add `any_target` union member to value payload
Both ANY-SERIES! and ANY-WORD! will hold binding information that is
either relative or specific.  The most efficient way to process this
information is to have it at the same location in the value payload
structure for both types.

This formalizes the relationship and the C specification clause that
allows reading the common member.  The rule seems that it may not also
apply to writing, so writes likely must be conditional on the word
or array payload (unless they become "Reb_Word_Or_Array", which can
be left for the future to decide if that is necessary)