INIT_WORD_INDEX_Debug function
Pursuant to the desire to put more debugging code in functions instead
of ever-growing macros, this updates the logic from the INIT_WORD_INDEX
macro and puts it into a function.  It checks both relative and
specific bindings for being in range at the time of assignment, and
that the symbols match the symbol of the word.