Separate macro for series initialization
Originally macros like `VAL_SERIES(value)` would provide direct LVALUE
references to the field in the REBVAL.  A trend has been rising that
tends to not do that, because there's more checking or otherwise that
needs to be done in the macro.  C cannot take a comma list like
`(assert(something), lvalue) = 10;` and allow that to achive an
assignment of 10 to `lvalue`.

There was a hack to continue allowing VAL_SERIES(value) to be an lvalue
by calling a debug routine that returned a pointer to a series, and
then assigning that.  e.g. it's legal to write:

     *assert_and_return_pointer(lvalue) = 10;

This removes that hack and goes ahead and makes it so that VAL_SERIES
has a separate INIT_VAL_SERIES to achieve the lvalue result.  VAL_INDEX
is still an lvalue, however.