REBVAL/RELVAL type checking up through %c-path.c
The compilation proceeds in alphabetical order.  These relatively light
RELVAL/REBVAL type checking changes are enough to get up to %c-path.c
with the type checking enabled.  Then the checks are re-disabled to
do a build and run it through the tests.

(While any particular set of these commits is fairly arbitrary, doing
it in steps helps with locating where a regression might have come from
in git bisection.)