Assert_REBVAL_Writable => Assert_Cell_Writable
Now that the writability check applies to RELVAL and REBVAL, the
explicit naming of REBVAL seems inappropriate.  Rather than change it
to the specific term RELVAL this calls it a cell.

It also re-instates the alignment assert, which may be the sort of thing
that is only done in an exhaustive build--but important to know if any
of the REBVALs aren't getting aligned on 32 or 64-bit boundaries.