Reb_Frame tweaks: label_sym, do_count
This takes the "opt_" off the front of label_sym, as it was kind of
noisy...and other fields are "opt" (like the ->func field, only valid
if the function is running).  The idea of needing to have the Do
wrappers pass in a symbol for APPLY became unnecessary, which was the
initial motivation behind stressing the optionality.

It also changes do_count to be stored as a REBUPT just to make the
alignment easier, e.g. for RenCpp.