Move variadic DO to r3-legacy, allow 0-arity DO
The warning that DO was not allowed to run functions was based on the
idea that core functions should not rely upon being variadic for their
function--and that this made it impossible to wrap DO.

With the implementation of VARARGS! it is possible to write and wrap
variadics.  So this moves the C hack required to keep DO running
functions to a usermode implementation of r3-legacy.  It also brings
a welcome reduction to the overall length of the evaluator code.

In addition, this re-enables DO for zero-arity functions.