Move state balancing check into XXX_Checks_Debug
Initially the state snapshot was not part of the Reb_Frame, due to
wanting to keep it small.  But since it is only in the debug build, it
really makes no difference...and it had to be moved into the frames
in order to debug check "continuations" (only implemented in the case
of APPLY, at this time).

With the state in the frame, it's possible to move the actual check
into the debug-only routines and out of the main evaluator function.
(Which it's better to make shorter if at all possible.)

Also improves the messages delivered for DO_COUNT_BREAKPOINT, and makes
the feature more obvious.  Now the breakpoint will place the debugger
directly on the switch statement when it happens.