Comment legacy extension code, clarify frame naming
Comments to explain what the RXIARG and COMMAND! mechanism is for, that
it is deprecated, and also tries to name variables to more clearly
differentiate a Reb_Frame structure from the "RXIFRM"... to help
eliminate confusion, as one is proxied to the other.  Hence an RXIFRM
will not be called a "frame" or "frm", but always an "rxifrm".

(Which is fine, as it won't exist indefinitely, we'll hope.)