Kill %reb-value.h and %reb-series.h
This gets rid of two parallel struct definitions of the internal
Reb_Series and Reb_Value types.  These become inconsistent with the
actual definitions as they change, because they are not included in
common by the core (which uses sys-value.h and sys-series.h instead)

It turns out these do not have anything to do with RXIARG at all, and
have what appears to be only one usage in R3-View...which is easy to
replace with an RL_XXX call.