Add PRINT/ONLY, only dialect nested PRINTs if literal
This adds an /ONLY switch to PRINT in order to have it print a value
without adding a newline, nor reducing it if it's a block.  PRIN is
then rewritten in terms of PRINT/ONLY (reducing blocks first).

It also changes the nesting logic so that a block will only be
reduced and handled in the "print dialect" if it appears literally.
Hence `print ["a" [reverse "hello"] "b"]` will run the reverse, while
`foo: [reverse "hello"] print ["a" foo "b"]` will not.  This is a
step toward avoiding accidental reducing, although further changes to
the "print dialect" will likely be needed.