Comment Resolve_Path, make it work >1 level deep
Resolve_Path is a stripped down piece of Do_Path which seems to only
be used by the PROTECT code to figure out what object to protect a
key in.  It's a bit of a weak routine, but worse than that, it didn't
work for depths greater than 1 because it did not update the variable
as it went.

This slightly improves it, and makes it use a RELVAL for array access.