Fix macros with duplicating side-effects
Without inline functions, an ALL_CAPS macro in Ren-C should not have any
parameters passed to it with side-effects, because the expression may
be repeated.

The INIT_VAL_SERIES() and INIT_VAL_ARRAY() macros were introduced and
added checks, but did not make sure the parameters were side-effect
free.  This fixes that problem in these cases.