"Closure"-semantics don't require rebind phase
Historically it was required on each call that closures do a full copy
of their body, and then as a second pass rebind the words to an object
to hold the function's locals and args specifically (and "durably").

This commit progresses specific binding to the point of being able
to do a one-pass deep copy of the body, in a way that derelativizes
according to the frame.  This is the same process that would be used
to derelativize copies made by the user, and requires no rebinding step.

Although a large part of the codebase has not been updated to correctly
"thread" the specific context through calls in and out of the core,
this is able to work accurately because it is done in one step.  Other
parts of the system are able to interact with specific binding at this
time due to tolerance and fallback on dynamic binding, which could
introduce "erratic" binding for cases of leaked locals from ordinary
FUNCs that aren't using <durable> ("closure") semantics.