Give specifier temp storage in PVS path state
Path dispatch has always seemed fairly dodgy.  The desire to extend
it to be able to pick at a "sub-value" granularity in the dispatchers
for GOB! and STRUCT! led to calling Next_Path() from within the
dispatchers themselves, which made it even more dodgy.

One problem was that Next_Path() may need to do an evaluation (such as
for a GROUP!) which means that the selector will be a generated value.
If a stack variable is used inside of Next_Path() to hold that generated
selector, then by the time it returns it will be invalid.  This was
a possibility with the sub-path dispatch (args/(1 + 1) vs. args/2) but
had apparently not come up in practice.

Modifications for relative values meant that relatively bound words
like `args/:rel` would also be examples of needing generated values,
as the bit pattern of a fully specified word would be different.  This
led to creating the same circumstance that a GROUP! would have made.

As the PVS is a path state structure, the simple solution to this is
to move the temporary slot into the state.