Nix "no-arg" function call IS_END() optimization
Calling functions with no arguments is extremely rare...especially when
no-argument user-functions will typically have a definitional RETURN
or LEAVE anyway.  The benefit was tiny to noticing the exception to
the case, and more likely just an extra IS_END() check on the average
case of having at least one argument or local.