Fix up some parent/child specifier calls
This adjusts some parent and child specifier calls to their correct
semantics, as the specifier logic is starting to become more "real".
When digging into an ARRAY and then needing a specifier, the specifier
from the array should be used if the array is specific, or a parent
specifier from elsewhere if it is relative.

While making the changes, an issue in COMPOSE that would crop off the
beginning of nested non-block arrays was corrected.

Also includes some more progress on comment reduction in Do_Core().