Simplify series common actions, kill DECIDE
DECIDE was some superfluous macro abuse that would allow the creation
of a pair of goto labels, and then implicitly jump to one or the other
based on whether the condition it was passed was true.  It has been
slowly phased out but this takes care of the few remaining instances.

Also this makes it slightly clearer how the series common actions
are handled, and avoids mixing in -1 with the REB_R return results,
which may be changing in the future.