END_VALUE instead of NULL terminates variadic Error()
In C, NULL is the constant 0...which in variadics uses integer rules.
This means that reading it back using `va_arg` as a pointer type is
not legal.  Error() didn't require NULL since the error templates
signal how many arguments they take.  But it was allowing it as a way
to signal end of arguments.  This changes the optional signal to
END_VALUE and uses IS_END() as the test instead.