make SER_REST work for non-dynamic series
The SER_REST macro was giving back garbage for series using the trick
of being just one series node in size.  This usually didn't matter
since they were fixed size and could not expand, but it turns out even
so an assertion could be tripped in Valgrind by it.

This adjusts it to give an accurate answer using the same logic as
SER_LEN does for the non-dynamic case.