Fix definitional LEAVE/RETURNs type checking
This fixes a premature type check (and in PROC's case, value-
suppression) which should only be run if a THROW, RETURN, or LEAVE is
not crossing out of a function with a definitional LEAVE or RETURN.
By running before the checks for throws, it was wiping out the thrown
value or type checking it when it shouldn't.

This moves the type check after a point where the throws have GOTO'd.