More rigorously enforce label_sym rule for frames
While previously a missing or incorrect symbol on a stack frame was
likely to go unnoticed, the debugger makes any errors much more apparent
(and also, debugging the debugger when there are symbol errors is
doubly frustrating).  This adds assertions and macros to make it
more rigorous that a frame will always have SYM_0 when not running a
function, and always have a non-SYM_0 value when a function is running
or gathering arguments.