Add RUNNING? and PENDING? tests for FRAME!
This gives an API ability to tell if a frame is either in mid-execution
or in the process of gathering its arguments prior to execution.

Note: The current implementation of the "accessible" flag for contexts
is that the flag is only valid to be checked for stack-based contexts.
But the running? and pending? routines were reading it on all contexts.

(It's probably better to switch to making it an "inaccessible"
flag instead, so that it's not necessary to check if something is
stack based...although that would mean the same flag bit couldn't
be used for another meaning in non-stack contexts.)