64-bit alignment fixes
The x86 family is tolerant of "bad" alignments for 64-bit quanitites on
32-bit platforms (such as a `double` being 8 bytes long, but placed
on an address where `address % 8 != 0`).  The emscripten JavaScript
build has some problems with this.

Bad 64-bit alignments were happening in Rebol caused by two factors.

One was the usage of a non-standard #pragma for "byte packing"
the structures related to REBVAL and REBVAL payloads.  By requesting
the compiler to pack at 4 byte boundaries, it subverted the natural
padding which would be inserted in structs and unions that would
ensure their alignment was correct.  The pragma is necessary to get
128-bit REBVALs on 32-bit platforms instead of 150-bit ones, so this
keeps it but properly pads the 64-bit values in payloads relative to
the start pointer of the value.

The other problem was that the pooled allocator Make_Node() was giving
back memory that only guaranteed 32-bit alignment.  While malloc()'s
guarantee is that it returns memory "suitably aligned for any type"
(compiler-defined), Rebol does not use types with alignments larger
than 64-bit at this time.  So this ensures that Make_Node() returns
memory that is always 64-bit aligned.  It does this by having data
types which are used in pools be multiples of 8 bytes on both 32-bit
and 64-bit builds.

This also gets rid of the "unteger" field inside of REBVAL.  This was a
trick used mainly by the SHIFT native in order to write a signed integer
into a REBVAL payload and then read it out as an unsigned value.  That
causes undefined behavior, as unions offer no such guarantee for when
the types are different.