Document or eliminate integer/decimal aliasing
While implementing the padding for 64-bit alignment, a temporary
mismatch of the positions used for 64-bit integers and 64-bit floating
point numbers exposed that there were several cases where VAL_INT64
was being used as a method for accessing the raw bits of a value that
was actually a REB_PERCENT or a REB_DECIMAL.  There was also a bug
in comparisons coming from the 64-bit decimal number being used for
REB_MONEY without a conversion.

This makes the debug build check that VAL_INTEGER() is used only on
REB_INTEGER, and VAL_DECIMAL() is used only on REB_DECIMAL and
REB_PERCENT.  It also formalizes the access of the 'raw bits' via an
integer cast as VAL_DECIMAL_BITS() to make it clearer what is going
on.  (The idea itself of casting to a 64-bit integer should be
reviewed, as it may be undefined behavior...but this just points it
out for now.)