Retire varless ("frameless") native optimization
The concept of the "frameless" optimization was that pushing stack
space, iterating argument lists, and type-checking could be foregone
or simplified in certain foundational operations (IF, EITHER, QUOTE,
CASE) by having those routines operate directly on the input series
itself--as if those natives were part of the Do_Core evaluator itself.

Between the time of the initial idea and today, some changes made the
idea less of a win.  The evaluator had its state turned "inside out"
such that all functions had frame objects, and it was merely that these
routines were becoming "varless".  Also, the number of conditions under
which varlessness could be exercised were reduced, so several checks
were needed to decide whether to put it into effect--creating cost for
both varless and non-varless routines.

The debugger in particular exposed complexities, because a varless
native could not have its parameters introspected in the stack.  Error
reporting had to take varlessness into account, in terms of whether
the error should be reported in the caller (in the case of a failed
type check) or in the varless native (if the code was part of the
implementation post-parameter gathering.

In order to better face irreducible complexities (such as specific
binding), this eliminates the setting for natives.  However, the API
which was developed to support them has been renamed from the "varless"
API, and is now the "low-level DO API"--which is still in use, and
helped in the design of other features.