Switch Make_Frame to fill with void, not |
With voids no longer "value types", they are free to represent other
states.  With FRAME!, there are two possibilities for what it can
mean: It's either the absence of an <opt> parameter, or if doing
"specialization" then it's an unspecialized argument which should
be gathered at the callsite.  It is much more suitable than the
"real" value type BAR! was for this purpose, so it is replaced.

This is an incremental update of the way specialized frames were
initially called in the evaluator (prior to the SPECIALIZE native).
It eliminates the ET_FRAME evaluation type, and will only do
specialization on function calls.  EVAL can no longer be used to
interpret a frame as a specialization, and DO is used to invoke a
FRAME! value using the same semantics as apply (voids are <opt>)

The code is rough, but better than it was.  It has some more things
to integrate into it from the specific-binding rebase before
working on it too much.