Make lone underscore literal "BLANK!/NONE!"
This change makes a single _ equivalent to a literal NONE! value, and
is how a NONE! is molded to the display.  It establishes BLANK! as an
alias, and a test function BLANK?...but with no word BLANK to look
up to the value, as the underscore is intended to speak for itself.

WORD! with an underscore-only name would have to be done in construction
syntax.  (Note: #_ was seen in use in a very old copy of Red.)

The rationale for introducing this new literal form is due to an interest
in making it more useful in dialecting, and to remove it from the set of
things that the users might treat as a variable.  This will make it a
safer choice for dialects which just need to indicate a "slot", but do
not wish to establish their own keyword for slots.  Since NONE! needed
a literal form anyway, letting it do double duty as this shared idea
takes care of two things at once.

Notably, it is of use in the PRINT dialect for indicating a SPACE.
Attempts to come up with a satisfying solution for print which auto-spaces
has generally failed, as it is conceptually flawed to come up with ways
of putting in indicators to say "before there was something here, and now
there is nothing".  By making NONE!'s (blank's!) literal form a very
easy to type and regard symbol for indicating a space, it makes it less
painful, e.g.:

     print ["The value is" x] => print ["The value is" _ x]

While that reduced case may make it appear that the underscore is "ugly",
the "cleanliness" of the left case is generally an illusion.  More often
than not, code has to work its way out of the trap of the automatic
spacing.  Combined with the use of BAR! for newline, this presents
pleasing answers for PRINT.

Changes to PRINT will be in another commit.