no evaluate MAKE OBJECT!, start object specs
This begins to introduce the notion of an object spec.  Creating an
object with MAKE OBJECT! now requires a spec and a body, such as:

     make object! [
         [a b c]
         [a: 10 c: 20]
     ]

(There is a compatibility layer in r3-legacy, currently enabled by
default, for the old behavior if the definition does not start with
a block value.)

This is part of a more complex spec language, which would allow things
like type-constraining the fields (in the same way that functions can
specify types for their parameters) or express other properties.

But the near term pressure on developing this format is to find a way
to represent objects that works for any field type (e.g. if the value
is a SET-WORD! or a GROUP!), as well as to be able to have values that
are present in the spec but abesent from the values.  This means there
is a representation for an "unset" that does not require a physical
assignment to take place.  For instance, in R3-Alpha this would not
have been legal:

    make object! [a: 10 b: #[unset] c: 20]

CONSTRUCT is repurposed as the arity-2 object creation routine (spec
and body), while HAS takes the arity-1 job.  Both will automatically
scan the body for fields in the way MAKE OBJECT! previously did.  Both
will also do evaluations inside the body, unless /ONLY is used.

While this is very much a work in progress, it is operational enough to
run some of the testing projects...and is enough of a start to seek
the elimination of reified unsets.