sync Value_To_RXI() extern signature in %a-lib.c
Value_To_RXI was changed to take a pointer to an RXIARG as an `out`
argument vs. returning a value.  However, several routines that were
labeled as "local functions" in %f-extension.c were actually imported
by an extern in %a-lib.c, which still had the old function signature.

This syncs the signature and tries to add a bit better documentation
of what was being done there...which appears to be a way of avoiding
the %reb-lib.h dependency as an include for most of the core.