Eliminate reified UNSET! and datatype
In Rebol, the NONE! value represents a conditionally false slot,
used as a placeholder to indicate no other value is there.  This
"reified nil" does not trigger an error when used in conditionals,
and is returned by some operations such as FIND.

Rebol2 and R3-Alpha also had another "reified" form of "no value",
which was used as an implementation detail of representing the bit
pattern in an object for the value a key mapped to when no value
was present.  This second reified form would generate errors in
some places, such as when used in the condition of an IF statement.
It was known as the UNSET! type.

This commit de-refies "UNSET!", in the sense that it is no longer a
datatype.  TYPE-OF hence returns a *none value* in response to a
type query on an expression returning no value:

    >> none = type-of ()
    == true

The ramifications and discussions are too long for a commit message,
see the column for unset changes in the porting guide Trello:

https://trello.com/c/rmsTJueg

Note this formalizes the <opt> tag for optional arguments, which had
previously been working through a typeset alias for UNSET!.  It also
unifies the REB_TRASH as a modality of REB_UNSET in the debug build
(the main reason for the existence of a separate trash type was to not
expose the type to the users, and so if UNSET is already not meant to
be exposed that works just as well).