Eliminate the optional argument variadic trick
R3-Alpha supported a feature on quoted arguments, where if they were
both quoted and accepted UNSET! they would allow the argument to go
missing:

    >> foo: func [:x [unset! integer!]] [
        if unset? :x [print "This hack exists to make HELP etc. work"]
    ]

    >> (foo)
    This hack exists to make HELP etc. work

Effectively the function becomes a very limited form of variadic.  It
forced the function in question ot take its argument quoted, even if
that did not make sense.

Given the existence of full variadics in Ren-C, this hack does not make
sense to continue supporting.  This converts HELP to use a variadic
(as well as to use a PROCEDURE since it never returns a result)