Eliminate Get_Word_Name() wrapper
A problem in the emscripten build revealed there is a routine called
Get_Word_Name() that takes a REBVAL*, and if it is NULL subtly returns
the literal string "unnamed".  This behavior is undesirable, and clients
that need NULL tolerance are rare and should be coded specially to
handle the case.

Several uses were in commented out code.  This removes those, and
inlines the rest without the NULL tolerance.  Any attempts to get the
name of a NULL REBVAL* will now crash at the offending location.