Re-introduce legality of `x: ()` assignments
With the demise of "reified unsets", the role of the non-reified unset
(a "void" or "no expression result") is being reconsidered.  Its
advantage as a return value from routines that need an "out of band"
result, such as to distinguish `first []` from `first [_]`...is high.

Historically Rebol has sought to be ergonomic by conflating "no result"
with a "NONE! result", because of the hot-potato nature of handling an
UNSET! value.  A non-negotiable aspect of the hot-potato is on typical
GET, because otherwise a typo like `my-3-arg-funtcion 1 2 3` would
evaluate to 3 without an error.

However, the advantage of a "hot potato" on writes is less obvious.  It
provides locality of errors when an assignment would wind up clearing
a variable instead of setting it.  However, this often needs to be
subverted via SET/OPT (a.k.a. set/any), and with locals gathering
has to be `set/opt (quote item:) ...` instead of just `item: ...`

Given the hot-potato nature of reads, it's also one of the least helpful
cases.  Consider:

    str: some-function a b c
    if str = "something" [...]

In this case, should some-function return no value and str: accept that
as unsetting the str, it would still generate an error on the `str`
access.  Yet if some-function returned another "good" type like an
integer!, then it would not error and just fail quietly in the
comparison.

The ENSURE primitive exists to make sure a value is acquired:

    str: ensure some-function a b c

And there are perhaps other type-based constraints that could be
developed, so that only variables marked as legally optional would
allow an assignment from voids.  Such a feature could also be used to
constrain to other types, catching INTEGER! to STRING! for instance.

But for the general ergonomics of handling the soon-to-be-more-common
VOID? return results, being able to assign them easily is too important
with the concrete advantages of stopping the assignments too limited
in the scheme of things.