UNSET => VOID name change
In the continuing effort to clarify the internals and bring them into
alignment with the external, this gets rid of SET_UNSET(...) and other
confusing constructs by embracing VOID.

With this change, setness and unsetness is a property of variables in
contexts, and not of values.  Void values are prohibited in blocks but
are used as markers in contexts and maps for optional slots.  They can
also be transiently generated during evaluations.

To further drive home that VOID! is not a type, this makes the REB_XXX
for "no value" follow the pattern of other "not a value" types (like
SYM_0).  There will likely be a VAL_TYPE and VAL_TYPE_OPT distinction
for those routines that want to explicitly express that they are
willing to deal with void values...which would help tighten up the
invariant that there aren't supposed to be void values in blocks/etc.