Move SAME? out of Compare_Type code, drop EQUIV?
The SAME? function checks bindings, and this means that it's not
sufficient to compare relative values (RELVAL).  If a comparison needs
REBVAL instead of RELVAL, that means that a specifier must be used
in the recursion or REBVAL bits must be generated each time.

The CT_XXX functions had somewhat of a poor interface already (the
arbitrary integer `mode`) so this removes mode 3 for SAME?.  It also
drops mode 1 which was EQUIV?--which is not present in Red, and
represents an underspecified notion of comparison to be built into
the system vs. just writing it as a usermode function for those who
would need to use its semantics.

For now, if someone wants to check if words or blocks have some
common binding properties, that must be done manually.