MT_XXX incremental change, RELVAL known hacks
With the goal of being able to "flip the switch" such that RELVAL and
REBVAL are type-incompatible, this injects a lot of fake KNOWN()
macros into the %t-struct.c and %t-routine.c files, as a way of
marking them to be dealt with in the future.  After the change they
will work if given non-relative data, but assert if any relative
words are seen--and can be dealt with at that time.

This also makes one step toward a change for the MT make type routine
to take a specifier.  What will ultimately need to happen is that those
should take a specified REBVAL which corresponds to a make argument,
instead of a RELVAL* position that is incremented...and make should
stop being an action (A_MAKE) instead dispatching directly to that
routine.