Ground Zero boot with fully specific binding
This completely eliminates the dynamic binding code.  Hence the stack
is never consulted during word lookup...it is always using specifiers
threaded through the ANY-ARRAY! values.

These changes only get it up to the point of booting the system.  The
transition is systemic, and will take significantly more work to get
it to run reliably.  A large amount of debugging code should help to
catch mistakes as they are triggered.