Make BODY-OF work by inventing an expired frame
Generally speaking, user code is exposed to arrays and words inside of
a function body when those functions are running.  These instances
have a frame threaded in as a "specifier".

The BODY-OF reflector doesn't have a frame to use, however.  One
approach would be to create a special kind of array copier or copy
mode that does not need a specifier.  But the "cheap" technique used
here is just to make a dummy frame that matches the function, and
then say that frame is expired.  This means the body of the function
will act like it was specifically bound to a frame off the stack.

Also adds assertion that Make_Error is not given relative values.
It takes varargs, so many places still do--until some type checking
technique is added in the C++ build to stop it at the callsites.