Make FUNCTION variables indefinite extent
The argument for making arguments to FUNCTION indefinite extent is to
enable the natural coding construction:

    foo: function [x]
       y: 20
       return function [z] [x + y + z]
    ]

If `x` and `y` were not available after the function terminated, then
the returned function would be useless.

This flips the switch for FUNCTION but not for FUNC, which defaults to
indefinite extent.  Its strange name hints that it might be quirky,
and one of those quirks is that indefinite extent must be explicitly
requested or it is assumed to not be needed.

Consequently, CLOSURE and FUNCTION are now the same in Ren-C.