Make invalid type error take the type explicitly
This changes the invalid type error to just take the type.  This helps
make it usable for both RELVAL and REBVAL, in a way that doesn't
suggest a likely usage error to the reader--as passing RELVALs to
errors without a specifier is generally bad.

Part of patching the struct code to allow the %gtk.r sample to work.