Use Reb_Frame output for parse loop result
The Parse_Rules_Loop recursive routine had a somewhat unclear return
result invariant in R3-Alpha.  The "result" field of the parse state
was initialized to 0, and then could be set to either 1 or -1.  This
result was in addition to the position returned as the main return
value (which could also be THROWN_FLAG or NOT_FOUND).

This auxiliary result was essentially an instruction as to whether a
containing loop should be interrupted...via ACCEPT or REJECT.  To
shoehorn the parse process into the REBNATIVE model for debugging
reflection, this implements those additional flags as throws (similar
to how a BREAK or CONTINUE would be implemented for an ordinary loop).

A helper wrapper is introduced, which draws to light that most callers
were ignoring this flag.