Parse code renaming of `item` to `rule`
The parse code used "item" to refer to a single REBVAL-item's-worth of
data in the rule array (so `some ["a" | "b"]` would consider the word
SOME an "item" and then the block an item, and when recursing call
the strings and bar "item" as well.

Though the word "rule" is a bit loaded in the sense of perhaps
suggesting a larger grouping of elements, it's confusing to know if
you are dealing with an item that is part of the input data or part of
the rules.  So naming this element `rule` instead of `item` makes the
code more clear.